Downgrade libpng again
New libpng combined with new wxWidgets appears to cause unholy crashes,
probably due to having two conflicting copies of libpng linked into the
same binary.
This needs to be investigated and fixed properly at some point (does
Aegisub even really need libpng outside of wxWidgets? It's a dependency
of freetype and libass, but libass only needs it for its test programs
and freetype probably only needs it for bitmap fonts?), but since this
entire branch is a hack anyway I'll just downgrade it for now.